+++ src/voxelizer.h
@@ -18,7 +18,7 @@
 #include <queue>
 
 #include "commons.h"
-#include "threadPool.h"
+#include "thread_pool.h"
 #include "timer.h"
 
 using namespace std;
