+++ test/collision_checker.h
